Code restructure: 2 separate commands now reside in seprate files. (#1)
